# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/instance": {
            "get": {
                "description": "Get all exports",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "Get exports",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "exports",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lib.Instance"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/instance/{id}": {
            "delete": {
                "description": "Remove an export.",
                "produces": [
                    "application/json",
                    "text/plain"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "Delete export",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "export id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "207": {
                        "description": "Multi-Status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/async-doc": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documentation"
                ],
                "summary": "Get asyncapi",
                "responses": {
                    "200": {
                        "description": "asyncapi file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/databases": {
            "get": {
                "description": "List all export databases",
                "consumes": [
                    "application/json",
                    "text/plain"
                ],
                "produces": [
                    "application/json",
                    "text/plain"
                ],
                "tags": [
                    "Export Database"
                ],
                "summary": "Get databases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "deployment",
                        "name": "deployment",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "owner",
                        "name": "owner",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "databases",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lib.ExportDatabase"
                            }
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an export database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Export Database"
                ],
                "summary": "Create database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "export database data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.ExportDatabaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.ExportDatabase"
                        }
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/doc": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documentation"
                ],
                "summary": "Get swagger",
                "responses": {
                    "200": {
                        "description": "swagger file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/instances": {
            "delete": {
                "description": "Remove multiple exports.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "Delete exports",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "export ids",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "207": {
                        "description": "Multi-Status",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "lib.Instance": {
            "type": "object",
            "properties": {
                "applicationId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "database": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "entityName": {
                    "type": "string"
                },
                "exportDatabase": {
                    "$ref": "#/definitions/lib.ExportDatabase"
                },
                "exportDatabaseID": {
                    "type": "string"
                },
                "filter": {
                    "type": "string"
                },
                "filterType": {
                    "type": "string"
                },
                "generated": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "measurement": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "offset": {
                    "type": "string"
                },
                "rancherServiceId": {
                    "type": "string"
                },
                "serviceName": {
                    "type": "string"
                },
                "timePath": {
                    "type": "string"
                },
                "timePrecision": {
                    "type": "string"
                },
                "timestampFormat": {
                    "type": "string"
                },
                "timestampUnique": {
                    "type": "boolean"
                },
                "topic": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Value"
                    }
                }
            }
        },
        "lib.Response": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}